route, controller aluno

+++ app/Http/Controllers/Api/AlunoController.php
@@ -2,13 +2,19 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Models\Aluno; 
+use App\Models\Aluno;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-
+use Validator;
+use DB;
 
 class AlunoController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show', 'store', 'update', 'destroy']]);
+    }
 
     /**
      * Display a listing of the resource.
@@ -17,8 +23,15 @@ class AlunoController extends Controller
      */
     public function index()
     {
-        return Aluno::all(); 
+        $alunos = Aluno::all();
+
+        if ($alunos->isEmpty()) {
+            return response()->json(['msg' => 'Nenhum registro encontrado'], 404);
+        }
+
+        return response()->json($alunos, 200);
     }
+
 
     /**
      * Store a newly created resource in storage.
@@ -28,8 +41,20 @@ class AlunoController extends Controller
      */
     public function store(Request $request)
     {
-        $aluno = Aluno::create($request->all());
-        return response()->json($aluno, 201);
+        $data = $request->all();
+
+        $validator = Validator::make($data, [
+            'nome' => 'required',
+            'cgm' => 'required|unique:alunos'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+
+        $aluno = Aluno::create($data);
+
+        return response()->json(['msg' => 'Registro cadastrado com sucesso', 'data' => $aluno], 201);
     }
 
     /**
@@ -38,10 +63,17 @@ class AlunoController extends Controller
      * @param  \App\Models\Aluno  $aluno
      * @return \Illuminate\Http\Response
      */
-    public function show(Aluno $aluno)
+    public function show($id)
     {
-        return $aluno;
+        $aluno = Aluno::find($id);
+
+        if (!$aluno) {
+            return response()->json(['error' => 'Registro não encontrado!'], 404);
+        }
+
+        return response()->json($aluno, 200);
     }
+
 
     /**
      * Update the specified resource in storage.
@@ -50,10 +82,28 @@ class AlunoController extends Controller
      * @param  \App\Models\Aluno  $aluno
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Aluno $aluno)
+    public function update(Request $request, $id)
     {
-        $aluno->update($request->all());
-        return response()->json($aluno, 200);
+        $aluno = Aluno::find($id);
+
+        if (!$aluno) {
+            return response()->json(['error' => 'Registro não encontrado!'], 404);
+        }
+
+        $data = $request->all();
+
+        $validator = Validator::make($data, [
+            'nome' => 'required',
+            'cgm' => 'required|unique:alunos,cgm,' . $id
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+
+        $aluno->update($data);
+
+        return response()->json(['msg' => 'Registro atualizado com sucesso!', 'data' => $aluno], 200);
     }
 
     /**
@@ -62,10 +112,16 @@ class AlunoController extends Controller
      * @param  \App\Models\Aluno  $aluno
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Aluno $aluno)
+    public function destroy($id)
     {
-        $aluno->delete();
-        return response()->json(null, 204);
-    }
+        $aluno = Aluno::find($id);
 
+        if (!$aluno) {
+            return response()->json(['error' => 'Registro não encontrado!'], 404);
+        }
+
+        $aluno->delete();
+
+        return response()->json(['msg' => 'Registro removido com sucesso!'], 200);
+    }
 }
